backend/requirements.txt
beautifulsoup4==4.13.4
fastapi==0.116.1
langchain==0.3.27
langchain_community==0.3.27
langchain_openai==0.3.30
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
transformers==4.55.0
